simplify, break out common components
